no more redis, long live sqlite3
I made the decision to move off redis to sqlite3. Honestly most of the complexity in this project was around dealing with dates and "buckets" to make things efficient.. I don't really care THAT much about that right now and sqlite is plenty fast enough for the 1 user that actually uses this project.

It eliminates a bunch of complexity, and gives us the ability to easily add more functionality (users? hello!, pagination? yes please!) as I extend the project